requirements.txt
aiogram
python-dotenv
sqlite3
aiosqlite